+++ ai_snake_lab/constants/DEpsilon.py
@@ -8,7 +8,7 @@ constants/DEpsilon.py
     License: GPL 3.0
 """
 
-from utils.ConstGroup import ConstGroup
+from ai_snake_lab.utils.ConstGroup import ConstGroup
 
 
 class DEpsilon(ConstGroup):