+++ ai_snake_lab/constants/DModelLRNN.py
@@ -8,7 +8,7 @@ constants/DModelRNN.py
     License: GPL 3.0
 """
 
-from utils.ConstGroup import ConstGroup
+from ai_snake_lab.utils.ConstGroup import ConstGroup
 
 
 class DModelRNN(ConstGroup):