+++ ai_snake_lab/constants/DReplayMemory.py
@@ -8,7 +8,7 @@ constants/DReplayMemory.py
     License: GPL 3.0
 """
 
-from utils.ConstGroup import ConstGroup
+from ai_snake_lab.utils.ConstGroup import ConstGroup
 
 
 class MEM_TYPE(ConstGroup):